feat(passmenu): reworked passmenu
I ruined the elegance of the original passmenu, but I did add the
feature of being able to select the kind of field you want to copy, be
it password, username or email etc., based on the password
file structure convention as explained by the creator of `pass`.

Basically, it sends a list of options to dmenu, and stores the return
value in a variable. It then uses `awk` with ': ' (note the space) as a
field separator, and prints the value where the field matches the
variable.

It also spawns a subprocess that will empty the clipboard again. This
should be decently robust, as I took it largely from the original `pass`
source.