feat: rework `n` launch wrapper to spawn with an interactive shell
When called, it replaces the shell used to run this script with an `st`
process, that runs a new interactive shell, that runs an `nnn` process,
which is then subsumed by a shell process once finished. All this to
launch a terminal window with a specific cli application running, that
can continue as a regular terminal window once the cli application is
closed, without having dangling shells. This has the same result as
manually spawning a new terminal, and calling `n`, but can be called
from a hotkey or a launcher.